LayoutTests/http/tests/inspector/filesystem/read-directory.html

<!DOCTYPE html><meta charset="UTF-8">
<script src="../inspector-test.js"></script>
<script src="filesystem-test.js"></script>
<script>
document.addEventListener("DOMContentLoaded", runTest);
function test()
{
    var requestManager = new WebInspector.FileSystemRequestManager();

    InspectorTest.clearFileSystem(step1);

    function step1()
    {
        InspectorTest.createDirectory("/hoge", step2);
    }

    function step2()
    {
        InspectorTest.createDirectory("/hoge/fuga", step3);
    }

    function step3()
    {
        InspectorTest.createFile("/hoge/piyo.html", step4);
    }

    function step4()
    {
        requestManager.requestDirectoryContent("filesystem:http://127.0.0.1:8000/temporary/hoge", step5);
    }

    function step5(errorCode, entries)
    {
        InspectorTest.dumpReadDirectoryResult(errorCode, entries);
        requestManager.requestDirectoryContent("filesystem:http://127.0.0.1:8000/temporary/foo", step6);
    }

    function step6(errorCode, entries) {
        InspectorTest.dumpReadDirectoryResult(errorCode, entries);

        InspectorTest.clearFileSystem(step7);
    }

    function step7()
    {
        InspectorTest.completeTest();
    }
}

</script>
<p>Tests readDirectory.</p>
