LayoutTests/http/tests/inspector/filesystem/read-directory-expected.txt

Tests readDirectory.

errorCode: 0
entries:
  0:
    url: filesystem:http://127.0.0.1:8000/temporary/hoge/piyo.html
    name: piyo.html
    isDirectory: false
    mimeType: text/html
    resourceType: Document
  1:
    url: filesystem:http://127.0.0.1:8000/temporary/hoge/fuga
    name: fuga
    isDirectory: true
errorCode: 1
entries: (null)
